ocr-services/paddleocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paddlepaddle==2.6.2
paddleocr==2.7.3
Pillow==10.4.0
numpy==1.24.3
python-multipart==0.0.6
requests==2.32.3
